+++ src/program/Config.cpp
@@ -143,6 +143,8 @@ void Config::save(const std::string& gamepath) {
     settings.setValue("opengl_performance", sc.opengl_performance);
     settings.setValue("async_events", sc.async_events);
     settings.setValue("wait_timeout", sc.wait_timeout);
+    settings.setValue("game_specific_timing", sc.game_specific_timing);
+    settings.setValue("game_specific_sync", sc.game_specific_sync);
 
     settings.beginWriteArray("main_gettimes_threshold");
     for (int t=0; t<SharedConfig::TIMETYPE_NUMTRACKEDTYPES; t++) {
@@ -258,11 +260,13 @@ void Config::load(const std::string& gamepath) {
     sc.audio_channels = settings.value("audio_channels", sc.audio_channels).toInt();
     sc.audio_frequency = settings.value("audio_frequency", sc.audio_frequency).toInt();
     sc.audio_mute = settings.value("audio_mute", sc.audio_mute).toBool();
-    sc.audio_disabled = settings.value("audio_disabled", sc.audio_disabled).toBool();    
+    sc.audio_disabled = settings.value("audio_disabled", sc.audio_disabled).toBool();
     sc.locale = settings.value("locale", sc.locale).toInt();
     sc.virtual_steam = settings.value("virtual_steam", sc.virtual_steam).toBool();
     sc.async_events = settings.value("async_events", sc.async_events).toInt();
     sc.wait_timeout = settings.value("wait_timeout", sc.wait_timeout).toInt();
+    sc.game_specific_timing = settings.value("game_specific_timing", sc.game_specific_timing).toInt();
+    sc.game_specific_sync = settings.value("game_specific_sync", sc.game_specific_sync).toInt();
 
     sc.video_codec = settings.value("video_codec", sc.video_codec).toInt();
     sc.video_bitrate = settings.value("video_bitrate", sc.video_bitrate).toInt();