+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Option to disable audio completely (some games don't accept that and quit)
 * Hook LdrGetProcedureAddress() by patching ntdll.dll.so
 * Add a full wait setting
+* Add thread synchronization methods
+* Introduce game-specific settings
 
 ### Changed